Use CommonJS instead of custom transformers;
while we are left with non-pretty replacements of `require` with
`await require`, maybe those could be performed in an `after`
transformer in the future.

The advantage of this approach is in lower maintenance burden
(less to rewrite on each TypeScript major version bump) and
a working implementation of exports, over quite complex custom
transformer which was not fully functional for some edge cases.

The disadvantage is that the transpiled code looks less like
the original making debugging harder for plugin users.